+++ Source/WTF/wtf/FastMalloc.cpp
@@ -57,33 +57,6 @@
 
 namespace WTF {
 
-#if ASSERT_ENABLED
-thread_local static unsigned forbidMallocUseScopeCount;
-thread_local static unsigned disableMallocRestrictionScopeCount;
-
-ForbidMallocUseForCurrentThreadScope::ForbidMallocUseForCurrentThreadScope()
-{
-    ++forbidMallocUseScopeCount;
-}
-
-ForbidMallocUseForCurrentThreadScope::~ForbidMallocUseForCurrentThreadScope()
-{
-    ASSERT(forbidMallocUseScopeCount);
-    --forbidMallocUseScopeCount;
-}
-
-DisableMallocRestrictionsForCurrentThreadScope::DisableMallocRestrictionsForCurrentThreadScope()
-{
-    ++disableMallocRestrictionScopeCount;
-}
-
-DisableMallocRestrictionsForCurrentThreadScope::~DisableMallocRestrictionsForCurrentThreadScope()
-{
-    ASSERT(disableMallocRestrictionScopeCount);
-    --disableMallocRestrictionScopeCount;
-}
-#endif
-
 #if !defined(NDEBUG)
 namespace {
 // We do not use std::numeric_limits<size_t>::max() here due to the edge case in VC++.
@@ -232,14 +205,14 @@ void fastAlignedFree(void* p)
 TryMallocReturnValue tryFastMalloc(size_t n) 
 {
     FAIL_IF_EXCEEDS_LIMIT(n);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     return malloc(n);
 }
 
 void* fastMalloc(size_t n) 
 {
     ASSERT_IS_WITHIN_LIMIT(n);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = malloc(n);
     if (!result)
         CRASH();
@@ -266,14 +239,14 @@ TryMallocReturnValue tryFastZeroedMalloc(size_t n)
 TryMallocReturnValue tryFastCalloc(size_t n_elements, size_t element_size)
 {
     FAIL_IF_EXCEEDS_LIMIT(n_elements * element_size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     return calloc(n_elements, element_size);
 }
 
 void* fastCalloc(size_t n_elements, size_t element_size)
 {
     ASSERT_IS_WITHIN_LIMIT(n_elements * element_size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = calloc(n_elements, element_size);
     if (!result)
         CRASH();
@@ -289,7 +262,7 @@ void fastFree(void* p)
 void* fastRealloc(void* p, size_t n)
 {
     ASSERT_IS_WITHIN_LIMIT(n);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = realloc(p, n);
     if (!result)
         CRASH();
@@ -299,7 +272,7 @@ void* fastRealloc(void* p, size_t n)
 TryMallocReturnValue tryFastRealloc(void* p, size_t n)
 {
     FAIL_IF_EXCEEDS_LIMIT(n);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     return realloc(p, n);
 }
 
@@ -559,7 +532,7 @@ bool isFastMallocEnabled()
 void* fastMalloc(size_t size)
 {
     ASSERT_IS_WITHIN_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::malloc(size, bmalloc::CompactAllocationMode::NonCompact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -572,7 +545,7 @@ void* fastMalloc(size_t size)
 void* fastZeroedMalloc(size_t size)
 {
     ASSERT_IS_WITHIN_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::zeroedMalloc(size, bmalloc::CompactAllocationMode::NonCompact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -585,7 +558,7 @@ void* fastZeroedMalloc(size_t size)
 TryMallocReturnValue tryFastZeroedMalloc(size_t size)
 {
     FAIL_IF_EXCEEDS_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::tryZeroedMalloc(size, bmalloc::CompactAllocationMode::NonCompact);
     BPROFILE_TRY_ALLOCATION(NON_JS_CELL, result, size);
     return result;
@@ -606,7 +579,7 @@ void* fastCalloc(size_t numElements, size_t elementSize)
 void* fastRealloc(void* object, size_t size)
 {
     ASSERT_IS_WITHIN_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::realloc(object, size, bmalloc::CompactAllocationMode::NonCompact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -650,7 +623,7 @@ size_t fastMallocGoodSize(size_t size)
 void* fastAlignedMalloc(size_t alignment, size_t size)
 {
     ASSERT_IS_WITHIN_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::memalign(alignment, size, bmalloc::CompactAllocationMode::NonCompact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -663,7 +636,7 @@ void* fastAlignedMalloc(size_t alignment, size_t size)
 void* tryFastAlignedMalloc(size_t alignment, size_t size)
 {
     FAIL_IF_EXCEEDS_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::tryMemalign(alignment, size, bmalloc::CompactAllocationMode::NonCompact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -681,7 +654,7 @@ void fastAlignedFree(void* p)
 TryMallocReturnValue tryFastMalloc(size_t size)
 {
     FAIL_IF_EXCEEDS_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::tryMalloc(size, bmalloc::CompactAllocationMode::NonCompact);
     BPROFILE_TRY_ALLOCATION(NON_JS_CELL, result, size);
     return result;
@@ -700,7 +673,7 @@ TryMallocReturnValue tryFastCalloc(size_t numElements, size_t elementSize)
 TryMallocReturnValue tryFastRealloc(void* object, size_t newSize)
 {
     FAIL_IF_EXCEEDS_LIMIT(newSize);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::tryRealloc(object, newSize, bmalloc::CompactAllocationMode::NonCompact);
     BPROFILE_TRY_ALLOCATION(NON_JS_CELL, result, newSize);
     return result;
@@ -709,7 +682,7 @@ TryMallocReturnValue tryFastRealloc(void* object, size_t newSize)
 void* fastCompactMalloc(size_t size)
 {
     ASSERT_IS_WITHIN_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::malloc(size, bmalloc::CompactAllocationMode::Compact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -722,7 +695,7 @@ void* fastCompactMalloc(size_t size)
 void* fastCompactZeroedMalloc(size_t size)
 {
     ASSERT_IS_WITHIN_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::zeroedMalloc(size, bmalloc::CompactAllocationMode::Compact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -735,7 +708,7 @@ void* fastCompactZeroedMalloc(size_t size)
 TryMallocReturnValue tryFastCompactZeroedMalloc(size_t size)
 {
     FAIL_IF_EXCEEDS_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     return bmalloc::api::tryZeroedMalloc(size, bmalloc::CompactAllocationMode::Compact);
 }
 
@@ -754,7 +727,7 @@ void* fastCompactCalloc(size_t numElements, size_t elementSize)
 void* fastCompactRealloc(void* object, size_t size)
 {
     ASSERT_IS_WITHIN_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::realloc(object, size, bmalloc::CompactAllocationMode::Compact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -767,7 +740,7 @@ void* fastCompactRealloc(void* object, size_t size)
 void* fastCompactAlignedMalloc(size_t alignment, size_t size)
 {
     ASSERT_IS_WITHIN_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::memalign(alignment, size, bmalloc::CompactAllocationMode::Compact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -780,7 +753,7 @@ void* fastCompactAlignedMalloc(size_t alignment, size_t size)
 void* tryFastCompactAlignedMalloc(size_t alignment, size_t size)
 {
     FAIL_IF_EXCEEDS_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::tryMemalign(alignment, size, bmalloc::CompactAllocationMode::Compact);
 #if ENABLE(MALLOC_HEAP_BREAKDOWN) && TRACK_MALLOC_CALLSTACK
     if (!AvoidRecordingScope::avoidRecordingCount())
@@ -793,7 +766,7 @@ void* tryFastCompactAlignedMalloc(size_t alignment, size_t size)
 TryMallocReturnValue tryFastCompactMalloc(size_t size)
 {
     FAIL_IF_EXCEEDS_LIMIT(size);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::tryMalloc(size, bmalloc::CompactAllocationMode::Compact);
     BPROFILE_ALLOCATION(COMPACTIBLE, result, size);
     return result;
@@ -812,7 +785,7 @@ TryMallocReturnValue tryFastCompactCalloc(size_t numElements, size_t elementSize
 TryMallocReturnValue tryFastCompactRealloc(void* object, size_t newSize)
 {
     FAIL_IF_EXCEEDS_LIMIT(newSize);
-    ASSERT(!forbidMallocUseScopeCount || disableMallocRestrictionScopeCount);
+    assertMallocRestrictionForCurrentThreadScope();
     void* result = bmalloc::api::tryRealloc(object, newSize, bmalloc::CompactAllocationMode::Compact);
     BPROFILE_ALLOCATION(COMPACTIBLE, result, newSize);
     return result;
